requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
boto3==1.14.63
botocore==1.17.63
cached-property==1.5.2
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
class-registry==2.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
django-filter==2.4.0
django-storages==1.10.1
djangorestframework==3.11.1
docutils==0.15.2
drf-yasg==1.17.1
filters==1.3.2
gunicorn==20.0.4
idna==3.2
importlib-metadata==4.6.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
kombu==5.1.0
MarkupSafe==2.0.1
packaging==21.0
Pillow==8.3.1
prompt-toolkit==3.0.19
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2020.1
redis==3.5.3
regex==2021.8.3
requests==2.26.0
ruamel.yaml==0.17.11
ruamel.yaml.clib==0.2.6
s3transfer==0.3.7
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
zipp==3.5.0
